// apps/admin/src/generated/api/schemas/listUserSessionsParams.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { ListUserSessionsStatus } from "./listUserSessionsStatus";
import type { Object } from "./object";

export type ListUserSessionsParams = {
  /**
   * @minimum 1
   */
  page?: Object;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: Object;
  search?: string;
  /**
   * Sort by one of: createdAt, lastUsedAt, expiresAt
   */
  sort?: string;
  status?: ListUserSessionsStatus;
  userId?: string;
  ipAddress?: string;
  dateFrom?: string;
  dateTo?: string;
};